Move ShowListViewController into its own file
